moved translation stuff to translation.py
